walk: gazebo: second attempt at a walking script that is deprecated
Second version of walking script, where the forward kinematics with the actual robot joints
from the simulation is done once and subsequent calculations are done from the previous
inverse kinematics results. It is being deprecated due to the robot having one foot in
the air still when it starts to lift the other foot.